% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcosTS.R
\name{arcosTS}
\alias{arcosTS}
\title{Helper of the arcosTS class}
\usage{
arcosTS(
  dt,
  colPos = "x",
  colMeas = NULL,
  col = list(Frame = "f", IDobj = "id", RT = NULL, IDcoll = NULL),
  interVal = NULL,
  interType = c("fixed")
)
}
\arguments{
\item{dt}{a data.table with time series in the long format.}

\item{colPos}{a vector with names of positional columns}

\item{col}{a list with names of other columns, i.e. list(Frame = , IDobj = , RT = , IDcoll = ), with names of the frame, object id, real time, and collective id columns.}

\item{interVal}{a numeric with the interval length.}

\item{interType}{a string to designate whether the time series has fixed or variable intervals; possible values fixed or var.}
}
\value{
an arcosTS object.
}
\description{
Helper of the arcosTS class
}
\examples{
cat("no examples")
}
