% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrendTS.R
\name{binMeas}
\alias{binMeas}
\alias{binMeas.arcosTS}
\title{"Smooth, de-trend, and binarise the measurement"}
\usage{
binMeas(
  obj,
  smoothK = 3L,
  biasK = 51L,
  peakThr = 0.2,
  biasMet = c("runmed", "lm", "none"),
  binThr = 0.5
)

\method{binMeas}{arcosTS}(
  obj,
  smoothK = 3L,
  biasK = 51L,
  peakThr = 0.2,
  biasMet = c("runmed", "lm", "none"),
  binThr = 0.5
)
}
\arguments{
\item{obj}{an arcosTS object.}

\item{smoothK}{an integer, length of the short-term median filter, i.e. smoothing, default 3L.}

\item{biasK}{an integer, length of the long-term median filter, i.e. de-trending, default 51L.}

\item{peakThr}{a double, threshold for peak detection from signal rescaled to [0,1], default 0.2.}

\item{biasMet}{method for de-trending, choose from runmed (median filter), lm (linear regression), none, default runmed.}

\item{binThr}{a double, threshold for signal binarisation, default 0.5.}
}
\value{
an arcosTS object.
}
\description{
Smooth, de-trend, and binarise the measurement
}
\details{
A wrapper for \code{detrendTS}.
}
\examples{
cat("no examples")
}
