% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binMeasLISA.R
\name{binMeasLISA}
\alias{binMeasLISA}
\alias{binMeasLISA.arcosTS}
\title{"Binarise the measurement using LISAs"}
\usage{
binMeasLISA(
  obj,
  d1 = 1,
  d2 = 10,
  statistic = c("localmoran", "localgeary", "localG", "localG*"),
  binThr = 0.5,
  binThrAsFrac = F
)

\method{binMeasLISA}{arcosTS}(
  obj,
  d1 = 1,
  d2 = 10,
  statistic = c("localmoran", "localgeary", "localG", "localG*"),
  binThr = 0.5,
  binThrAsFrac = F
)
}
\arguments{
\item{obj}{an arcosTS object.}

\item{d1}{numeric lower bound of local distance; default 0.}

\item{d2}{numeric upper bound of local distance; default 1.}

\item{statistic}{a character string specifying the LISA statistic that should be calculated. This can be one of "localmoran", "localgeary", "localG" or "localG*".}

\item{binThr}{a double, threshold for signal binarisation, default 0.5.}

\item{binThrAsFrac}{logical; if FALSE, the binThr is used as a fixed threshold to binarisa LISA; if TRUE, the binThr is treated as a fraction of LISA values independently per frame; default FALSE.}
}
\value{
an arcosTS object.
}
\description{
Binarise the measurement using local indicators of spatial association (LISAs) implemented in the package elsa.
}
\details{
Binarise the measurement using LISAs
}
\examples{
cat("no examples")
}
