% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\name{buyer_list}
\alias{buyer_list}
\title{Get list of business types listed in the BUYER_BUS_ACT in the ARCOS database}
\usage{
buyer_list(key = "WaPo")
}
\arguments{
\item{key}{Key needed to make query successful}
}
\description{
Get list of business types listed in the BUYER_BUS_ACT in the ARCOS database
}
\examples{
\donttest{
library(arcos)

bl <- buyer_list(key="WaPo")

head(bl)
}
}
\seealso{
\url{https://www.washingtonpost.com/graphics/2019/investigations/dea-pain-pill-database/#download-resources}

Other raw data: \code{\link{county_list}},
  \code{\link{county_raw_fips}}, \code{\link{county_raw}},
  \code{\link{drug_county_biz}},
  \code{\link{drug_fips_biz}}, \code{\link{drug_list}},
  \code{\link{pharmacy_raw}}, \code{\link{raw_data}}
}
\concept{raw data}
