% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFracActInColl.R
\name{calcFracActInColl}
\alias{calcFracActInColl}
\title{"Shuffle coordinates"}
\usage{
calcFracActInColl(objFull, objColl)
}
\arguments{
\item{objFull}{an arcosTS object with full data, without collective events.}

\item{objColl}{an arcosTS object with collective events. Can be also an arcosTS object returned from bootstrapping trackCollBoot function.}
}
\value{
a data.table with the fraction. If the objColl parameter comes from bootstrapping, fractions are calculated per bootstrapping iteration.
}
\description{
Calculate the fraction of activity in collective events
}
\details{
Calculate the fraction of total binarised activity involved in collective events.
}
\examples{
cat("no examples")

}
