% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowthColl.R
\name{calcGrowthColl}
\alias{calcGrowthColl}
\alias{calcGrowthColl.arcosTS}
\title{"Calculate collective events' growth"}
\usage{
calcGrowthColl(obj)

\method{calcGrowthColl}{arcosTS}(obj)
}
\arguments{
\item{obj}{an arcosTS object.}
}
\value{
a data.table with the width and height of the minimal directed bounding box, and the number of points used for the calculation.
}
\description{
Calculate the growth of collective events over time
}
\details{
Wrapper for the \code{getMinBBox2D} function that implements the rotating callipers algorithm. Calculates the minimal directed bounding box for every event at every time frame.
The longer side is taken as the "diameter". Currently works only with 2D data!
}
\examples{
cat("no examples")
}
