% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binMeasLISA.R
\name{calcLisa2D}
\alias{calcLisa2D}
\title{Calculate LISA in 2D}
\usage{
calcLisa2D(x, y, m, d1 = 0, d2 = 1, statistic = "localmoran")
}
\arguments{
\item{x}{a numeric vector with x coordinates.}

\item{y}{a numeric vector with y coordinates.}

\item{m}{a numeric vector with values at x and y.}

\item{d1}{numeric lower bound of local distance; default 0.}

\item{d2}{numeric upper bound of local distance; default 1.}

\item{statistic}{a character string specifying the LISA statistic that should be calculated. This can be one of "localmoran", "localgeary", "localG" or "localG*".}
}
\value{
a numeric vector with LISA statistic.
}
\description{
Calculate local indicators of spatial associations (LISAa) for a 2D point process.
Uses LISA implementations in the package elsa.
}
\examples{
library(ARCOS)
calcLisa2D(x = runif(1000),
           y = runif(1000),
           m = rnorm(1000))
}
