% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMoran.R
\name{calcMoran}
\alias{calcMoran}
\title{Global Moran's I spatial correlation}
\usage{
calcMoran(x, scol, mcol, ...)
}
\arguments{
\item{x}{a numeric matrix, with rows corresponding to objects' positions and columns as spatial coordinates.}

\item{scol}{a numeric scalar or vector that prescribes column numbers with spatial coordinates.}

\item{mcol}{a numeric scalar that prescribes a column number with the measurement.}

\item{...}{additional parameters passed to ape::Moran.I}
}
\value{
a list with observed, expected, sd, and p-value
}
\description{
Uses function Moran.I from the ape package to calculate the global spatial correlation of a point process with weights.
}
\examples{
library(ARCOS)
ARCOS::calcMoran(cbind(runif(100), runif(100), rnorm(100)), scol = c(1,2), mcol = 3)
}
