% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNNdists.R
\name{calcNNdists}
\alias{calcNNdists}
\alias{calcNNdists.arcosTS}
\title{"Calculate NN distances"}
\usage{
calcNNdists(obj, nnn = 1, act = FALSE)

\method{calcNNdists}{arcosTS}(obj, nnn = 1, act = FALSE)
}
\arguments{
\item{obj}{an arcosTS object.}

\item{nnn}{an integer, sets the n-th nearest neighbour; default 1.}

\item{act}{logical, if TRUE, the NN distance is calculated only on active objects, i.e. those with binarised measurement not smaller than 1; default FALSE.}
}
\value{
a data.table
}
\description{
Calculate nearest-neighbour distances
}
\details{
Wrapper for the \code{calcVecNNdists} function.
}
\examples{
library(ARCOS)
library(data.table)
ts = data.table(frame = c(1, 1, 2, 2, 2, 3, 3, 3, 3, 4, 4, 4, 5, 5),
                objid = c(1, 2, 1, 2, 3, 1, 2, 3, 4, 1, 2, 4, 1, 4),
                x = c(1, 3, 1.2, 2.5, 3.5, 0.9, 2.6, 2.9, 3.2, 1.1, 2.8, 3.1, 1, 3))

arcosTS(ts,
        colPos = "x",
        col = list(Frame = "frame",
                   IDobj = "objid",
                   RT = NULL,
                   IDcoll = NULL),
        interVal = 1.,
        interType = "fixed")

nnDists = calcNNdists(ts)
}
