% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPvalFromBS.R
\name{calcPvalFromBS}
\alias{calcPvalFromBS}
\title{Calculate p-value from bootstrapping}
\usage{
calcPvalFromBS(
  testStat,
  bsStat,
  corrected = FALSE,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{testStat}{a numeric, test statistic.}

\item{bsStat}{a numeric vector with statistic from bootstrapping.}

\item{corrected}{a logical; if TRUE, calculates a corrected p-value; default FALSE.}

\item{alternative}{either of "two.sided", "less", "greater".}
}
\value{
a scalar with the p-value
}
\description{
The corrected p-value calculate uses the formula from Davison and Hinkley (1997), Bootstrap Methods and their Application, p. 141.
}
\examples{
library(ARCOS)
ARCOS::calcPvalFromBS(2, rnorm(1000))
}
