% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsCollEvents.R
\name{calcStatsColl}
\alias{calcStatsColl}
\alias{calcStatsColl.arcosTS}
\title{"Stats of collective events"}
\usage{
calcStatsColl(obj)

\method{calcStatsColl}{arcosTS}(obj)
}
\arguments{
\item{obj}{an arcosTS object.}
}
\value{
an arcosTS object
}
\description{
Calculate statistics of collective events
}
\details{
Wrapper for the \code{calcStatsCollEvents} function to calculate statistics of collective events identified by the \code{trackCollEvents} function.
Calculates:
- event duration in time frames, \code{clDur},
- the total number of unique objects/cells that participate in an event, \code{totSz},
- the smallest number of objects/cells that comprise an event, \code{minSz},
- the largest number of objects/cells that comprise an event, \code{maxSz}.
}
\examples{
cat("no examples")
}
