% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsCollEvents.R
\name{calcStatsCollEvents}
\alias{calcStatsCollEvents}
\title{Calculate statistics of collective events}
\usage{
calcStatsCollEvents(
  inDTcoll,
  inCols = list(frame = "frame", id = "trackID", collid = "clTrackID", bootiter =
    "bootiter"),
  fromBoot = FALSE
)
}
\arguments{
\item{inDTcoll}{a data.table with collective events in the long format produced by the \code{trackCollEvents} function. Consists of 3 columns: integer frame number, object id, id's of collective events.}

\item{inCols}{a list with column names, \code{list(frame = , id = , collid = , bootiter =)}, that correspond to the integer frame number, object id, id of collective events, and bootstrapping iteration, respectively.}

\item{fromBoot}{logical, indicates whether input data comes from bootstrapping; default FALSE.}
}
\value{
a data.table with aggregated stats of collective events.
}
\description{
Calculate statistics of collective events identified by the \code{trackCollEvents} function.
Calculates:
- event duration in time frames, \code{clDur},
- the total number of unique objects/cells that participate in an event, \code{totSz},
- the smallest number of objects/cells that comprise an event, \code{minSz},
- the largest number of objects/cells that comprise an event, \code{maxSz}.
}
\examples{
cat("no examples")
}
\keyword{internal}
