% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNNdists.R
\name{calcVecNNdists}
\alias{calcVecNNdists}
\title{Calculate a vector of nearest neighbours}
\usage{
calcVecNNdists(x, k = 2, ...)
}
\arguments{
\item{x}{a numeric vector.}

\item{...}{parameters passed to the RANN:nn2() function.}
}
\value{
a vector with nearest neighbours
}
\description{
Calculate a vector of nearest neighbours
}
\examples{
library(ARCOS)



}
\keyword{internal}
