% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meas.R
\name{clipMeas}
\alias{clipMeas}
\alias{clipMeas.arcosTS}
\title{"Clip the measurement"}
\usage{
clipMeas(obj, clip, quant = FALSE)

\method{clipMeas}{arcosTS}(obj, clip, quant = FALSE)
}
\arguments{
\item{obj}{an arcosTS object.}

\item{clip}{a two-element vector with lower and upper measurement values (if quant is FALSE) or quantiles (if quant is TRUE) for clipping.}

\item{quant}{a logical to indicate whether the clip parameter is treated as lower and upper bounds of the measurement or as quantiles, default FALSE.}
}
\value{
an arcosTS object with a modified measurement column.
}
\description{
Clip the measurement
}
\examples{
cat("no examples")
}
