% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarized.R
\name{combined_buyer_annual}
\alias{combined_buyer_annual}
\title{Get annual total pills for each buyer (pharmacy, etc) in a county}
\usage{
combined_buyer_annual(county = "Mingo", state = "WV", key = "WaPo")
}
\arguments{
\item{county}{Filter the data to only this county (e.g. 'Mingo')}

\item{state}{Filter the data to county within this state (e.g. 'WV')}

\item{key}{Key needed to make query successful}
}
\description{
Get annual total pills for each buyer (pharmacy, etc) in a county
}
\examples{
\donttest{
library(arcos)

mingo_wv <- combined_buyer_annual(county = "Mingo", state="WV", key="WaPo")

head(mingo_wv)
}
}
\seealso{
\url{https://www.washingtonpost.com/graphics/2019/investigations/dea-pain-pill-database/#download-resources}

Other summarized: 
\code{\link{buyer_details}()},
\code{\link{combined_buyer_monthly}()},
\code{\link{summarized_county_annual}()},
\code{\link{summarized_county_monthly}()},
\code{\link{total_distributors_county}()},
\code{\link{total_distributors_state}()},
\code{\link{total_manufacturers_county}()},
\code{\link{total_manufacturers_state}()},
\code{\link{total_pharmacies_county}()},
\code{\link{total_pharmacies_state}()}
}
\concept{summarized}
