% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\name{county_list}
\alias{county_list}
\title{Get dataframe of counties, states, and fips codes that are represented in the ARCOS data}
\usage{
county_list(key = "WaPo")
}
\arguments{
\item{key}{Key needed to make query successful}
}
\description{
Get dataframe of counties, states, and fips codes that are represented in the ARCOS data
}
\examples{
\donttest{
library(arcos)

counties <- county_list(key="WaPo")

head(counties)
}
}
\seealso{
\url{https://www.washingtonpost.com/graphics/2019/investigations/dea-pain-pill-database/#download-resources}

Other raw data: 
\code{\link{buyer_list}()},
\code{\link{county_raw_fips}()},
\code{\link{county_raw}()},
\code{\link{drug_county_biz}()},
\code{\link{drug_fips_biz}()},
\code{\link{drug_list}()},
\code{\link{pharmacy_raw}()},
\code{\link{raw_data}()}
}
\concept{raw data}
