% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackCollEvents.R
\name{createCollEvents}
\alias{createCollEvents}
\title{Identify collective events from objects in the current frame}
\usage{
createCollEvents(
  inDT,
  inCols = list(x = "x", y = NULL, z = NULL, frame = "time", trackID = "trackID",
    clTrackID = "clTrackID"),
  inEps = 1,
  inMinPts = 1L,
  inClOffset = 0,
  DEB = T
)
}
\arguments{
\item{DEB}{}
}
\value{

}
\description{
Identify collective events from objects in the current frame
}
