% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\name{drug_list}
\alias{drug_list}
\title{Get list of drugs available in the ARCOS database}
\usage{
drug_list(key = "WaPo")
}
\arguments{
\item{key}{Key needed to make query successful}
}
\description{
Get list of drugs available in the ARCOS database
}
\examples{
\donttest{
library(arcos)

np <- drug_list(key="WaPo")

np
}
}
\seealso{
\url{https://www.washingtonpost.com/graphics/2019/investigations/dea-pain-pill-database/#download-resources}

Other raw data: 
\code{\link{buyer_list}()},
\code{\link{county_list}()},
\code{\link{county_raw_fips}()},
\code{\link{county_raw}()},
\code{\link{drug_county_biz}()},
\code{\link{drug_fips_biz}()},
\code{\link{pharmacy_raw}()},
\code{\link{raw_data}()}
}
\concept{raw data}
