% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export2napari.R
\name{export2napari}
\alias{export2napari}
\alias{export2napari.arcosTS}
\title{"Export data to Napari"}
\usage{
export2napari(objColl, dirpath, ...)

\method{export2napari}{arcosTS}(objColl, dirpath, ...)
}
\arguments{
\item{objColl}{an arcosTS object with collective events.}

\item{dirpath}{a string with the file name.}

\item{...}{parameters passed to (unused)}
}
\value{
writes files to a disk
}
\description{
Export data to Napari
}
\examples{
cat("no examples")
}
