% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genSynthMultiple2D}
\alias{genSynthMultiple2D}
\title{Random synthetic collective events in 2D}
\usage{
genSynthMultiple2D(
  nevents = 10L,
  maxt = 25L,
  maxx = 20L,
  maxy = 20L,
  maxdur = 5L,
  inSeed = NULL
)
}
\arguments{
\item{nevents}{an integer, defines the number of events; default 10.}

\item{maxt}{an integer, defines the maximum number of frames; default 25.}

\item{maxx}{an integer, defines the maximum width of the grid; default 20.}

\item{maxy}{an integer, defines the maximum height of the grid; default 20.}

\item{maxdur, }{an integer, defines the maximum duration of events; default 5.}

\item{inSeed}{an integer with the seed for the random number generator, default NULL.}
}
\value{
an arcosTS object
}
\description{
Create a sequence of collective events. A collective event is
a concentrically growing circle that increases its radius at every frame.
The location of the collective event is random on a `maxx`-by-`maxy` lattice,
and the duration is random between 1 and `maxdur` frames.
There are `nevents` that occur within `maxt` frames.
}
\examples{
library(ARCOS)
dts = genSynthMultiple2D()
}
