% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMinBBox}
\alias{getMinBBox}
\title{Rotating calipers algorithm}
\usage{
getMinBBox(xy)
}
\arguments{
\item{xy}{A matrix of xy values from which to calculate the minimum oriented bounding box.}
}
\value{
a list width numeric width and height of the bounding box
}
\description{
Calculates the minimum oriented bounding box using the
rotating calipers algorithm.
Credits go to Daniel Wollschlaeger <https://github.com/ramnathv>
The function modified from flightplanning-R package <https://github.com/caiohamamura/flightplanning-R/blob/master/R/utils.R>
}
\examples{
library(ARCOS)
getMinBBox(cbind(c(0,1,3,2,1), c(0,-1,1,3,2)))
}
