% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMinBBox2D}
\alias{getMinBBox2D}
\title{Rotating callipers algorithm in 2D}
\usage{
getMinBBox2D(xy, prec = 1e-08)
}
\arguments{
\item{xy}{a matrix of xy values from which to calculate the minimum oriented bounding box.}

\item{prec}{numerical, rounding precision; default 1e-08, lose to .Machine$double.eps^0.5.}
}
\value{
a list width numeric width and height of the bounding box
}
\description{
Calculates the minimum oriented bounding box using the
rotating callipers algorithm in 2D.
Credits go to Daniel Wollschlaeger.
The function modified from <http://dwoll.de/rexrepos/posts/diagBounding.html>.
}
\examples{
library(ARCOS)
getMinBBox(cbind(c(0,1,3,2,1), c(0,-1,1,3,2)))
}
