% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackLen.R
\name{histTrackLen}
\alias{histTrackLen}
\alias{histTrackLen.arcosTS}
\title{"Plot a histogram of track lengths}
\usage{
histTrackLen(obj, binwidth = NULL)

\method{histTrackLen}{arcosTS}(obj, binwidth = NULL)
}
\arguments{
\item{obj}{an arcosTS object.}
}
\value{
a ggplot2 object.
}
\description{
Plot a histogram of track lengths
}
\examples{
cat("no examples")
}
