% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateTS.R
\name{interpolateTS}
\alias{interpolateTS}
\title{Interpolate missing rows in time series}
\usage{
interpolateTS(inDT, inColID, inColFN, inColY, inFNfreq = 1L, inDeb = F)
}
\arguments{
\item{inDT}{a data.table with time series in the long format.}

\item{inColID}{a string with the name of the column with unique time series IDs.}

\item{inColFN}{a string with the name of the column with integer frame numbers.}

\item{inColY}{a string or a vector of strings with column names of numerical columns to interpolate.}

\item{inFNfreq}{an integer with the interval between frames.}

\item{inDeb}{logical, whether to output debug information.}
}
\value{
a data.table with interpolated missing time points.
}
\description{
Interpolate missing rows in time series
}
\examples{
library(ARCOS)
library(data.table)

dt = data.table(t = c(1,2,3,5,6,7),
                y = c(1,2,3,5,6,7),
                id = rep(1, 6))

dtInt = ARCOS:::interpolateTS(inDT = dt,
                      inColID = "id",
                      inColFN = "t",
                      inColY = "y",
                      inFNfreq = 1)

                      plot(dt$t, dt$y, type = "p")
                      points(dtInt$t, dtInt$y, col = "red", pch = 3)

}
\keyword{internal}
