% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateTS.R
\name{interpolMeas}
\alias{interpolMeas}
\alias{interpolMeas.arcosTS}
\title{Interpolate NAs and missing data}
\usage{
interpolMeas(obj)

\method{interpolMeas}{arcosTS}(obj)
}
\arguments{
\item{obj}{an arcosTS object.}
}
\value{
an arcosTS object.
}
\description{
Wrapper for the \code{interpolateTS} function.
}
\examples{
cat("no examples")
}
