% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcosTS.R
\name{is.arcosTS.fromBin}
\alias{is.arcosTS.fromBin}
\title{"Test whether the measurement in the arcosTS object has been binarised"}
\usage{
is.arcosTS.fromBin(x)
}
\arguments{
\item{x}{an object}
}
\value{
logical
}
\description{
Test whether an arcosTS object has a column with the binarised measurement calculated by the `binMeas` function.
}
\examples{
cat("no examples")
}
