% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcosTS.R
\name{is.arcosTS.fromBoot}
\alias{is.arcosTS.fromBoot}
\title{"Test whether an arcosTS comes from bootstrapping"}
\usage{
is.arcosTS.fromBoot(x)
}
\arguments{
\item{x}{an object}
}
\value{
logical
}
\description{
Test whether an arcosTS object comes from bootstrapping calculated by the `trackCollBoot` function.
}
\examples{
cat("no examples")
}
