% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcosTS.R
\name{is.arcosTS.fromColl}
\alias{is.arcosTS.fromColl}
\title{"Test whether an arcosTS object has collective events"}
\usage{
is.arcosTS.fromColl(x)
}
\arguments{
\item{x}{an object}
}
\value{
logical
}
\description{
Test whether an arcosTS object has collective events calculated by the `trackColl` function.
}
\examples{
cat("no examples")
}
