% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcosTS.R
\name{is.arcosTS}
\alias{is.arcosTS}
\title{"Test whether an object is an arcosTS object"}
\usage{
is.arcosTS(x)
}
\arguments{
\item{x}{object}
}
\description{
Test whether an object is an arcosTS object
}
\examples{
cat("no examples")
}
