% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keepSignifDig}
\alias{keepSignifDig}
\title{Keep significant digits in double numerical columns of a data.table}
\usage{
keepSignifDig(inDT, inDigits)
}
\arguments{
\item{inDT}{a data.table with time series in the long format.}

\item{inDigits}{an integer with the number of significant digits.}
}
\value{
a data.table with numeric columns trimmed to the provided number of significant digits.
}
\description{
Keep significant digits in double numerical columns of a data.table
}
\examples{
library(ARCOS)
library(data.table)

locdt = data.table(id = LETTERS[1:10],
                x = runif(10))

locdtTrim = ARCOS::keepSignifDig(locdt, 2)

}
