% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataFromFile.R
\name{loadDataFromFile}
\alias{loadDataFromFile}
\title{Title}
\usage{
loadDataFromFile(
  fname,
  colPos = "x",
  colMeas = NULL,
  col = list(Frame = "f", IDobj = "id", RT = NULL, IDcoll = NULL),
  interVal = 1L,
  interType = c("fixed"),
  ...
)
}
\arguments{
\item{path}{a string with path to images.}

\item{ext}{a string with pattern for image search/}
}
\value{
an arcosTS object.
}
\description{
Title
}
\examples{
cat("no examples")
}
