% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataFromImages.R
\name{loadDataFromImages}
\alias{loadDataFromImages}
\title{Title}
\usage{
loadDataFromImages(path, ext, thres = 0)
}
\arguments{
\item{path}{a string with path to images.}

\item{ext}{a string with pattern for image search.}

\item{thres}{a numeric with pixel threshold; default 0.}
}
\value{
an arcosTS object.
}
\description{
Title
}
\examples{
cat("no examples")
}
