% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{midPtCir}
\alias{midPtCir}
\title{Mid-Point Circle Drawing Algorithm}
\usage{
midPtCir(x0, y0, r)
}
\arguments{
\item{x0}{a numeric, defines the x coordinate of the circle's centre.}

\item{y0}{a numeric, defines the y coordinate of the circle's centre.}

\item{r}{a numeric, defines the circle's radius.}
}
\value{
a numeric matrix.
}
\description{
A utility function to create X/Y positions on a circle.
The algorithm adapted from <https://www.geeksforgeeks.org/mid-point-circle-drawing-algorithm/>
}
\examples{
library(ARCOS)
mcir = ARCOS:::midPtCir(1.4, 4.3, 5)
}
