% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{not_pharmacies}
\alias{not_pharmacies}
\title{Get list of misidentified pharmacies by BUYER_DEA_NOs}
\usage{
not_pharmacies(key = "WaPo")
}
\arguments{
\item{key}{Key needed to make query successful}
}
\description{
Get list of misidentified pharmacies by BUYER_DEA_NOs
}
\examples{
\dontrun{
library(arcos)

np <- not_pharmacies(key="WaPo")

head(np)
}
}
\seealso{
\url{https://www.washingtonpost.com/graphics/2019/investigations/dea-pain-pill-database/#download-resources}

Other geocoded data: \code{\link{buyer_addresses}},
  \code{\link{county_population}},
  \code{\link{pharm_cbsa}}, \code{\link{pharm_counties}},
  \code{\link{pharm_latlon}}, \code{\link{pharm_tracts}},
  \code{\link{reporter_addresses}},
  \code{\link{state_population}}
}
\concept{geocoded data}
