% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\name{pharmacy_raw}
\alias{pharmacy_raw}
\title{Download raw prescription data for specified pharmacy (by BUYER_DEA_NO) into R}
\usage{
pharmacy_raw(buyer_dea_no = "AB0454176", key = "WaPo")
}
\arguments{
\item{buyer_dea_no}{Filter the data to only this pharmacy (e.g. ‘AB0454176’)}

\item{key}{Key needed to make query successful}
}
\description{
Data from from non-contiguous states not yet processed and available.
}
\examples{
\donttest{
library(arcos)

solo_pharm <- pharmacy_raw(buyer_dea_no="AB0454176", key="WaPo")

head(solo_pharm)
}
}
\seealso{
\url{https://www.washingtonpost.com/graphics/2019/investigations/dea-pain-pill-database/#download-resources}

Other raw data: \code{\link{county_raw_fips}},
  \code{\link{county_raw}}, \code{\link{raw_data}}
}
\concept{raw data}
