% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.arcosTS}
\alias{plot.arcosTS}
\title{Plot an arcosTS object}
\usage{
\method{plot}{arcosTS}(obj)
}
\arguments{
\item{obj}{an arcosTS object.}
}
\value{

}
\description{
Plot an arcosTS object
}
\examples{
cat("no examples")
}
