% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTracks.R
\name{plotTracks}
\alias{plotTracks}
\alias{plotTracks.arcosTS}
\title{"Plot tracks"}
\usage{
plotTracks(obj, nmax = 20)

\method{plotTracks}{arcosTS}(obj, nmax = 20)
}
\arguments{
\item{obj}{an arcosTS object.}

\item{nmax}{an integer, maximum number of tracks to sample; default 20.}
}
\value{
a ggplot2 object.
}
\description{
Plot tracks from an arcosTS object
}
\details{
Plot a random selection of tracks in 1 or 2D.
}
\examples{
cat("no examples")
}
