% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randShiftMeas.R
\name{randShiftMeas}
\alias{randShiftMeas}
\alias{randShiftMeas.arcosTS}
\title{"Randomly shift the measurement"}
\usage{
randShiftMeas(obj)

\method{randShiftMeas}{arcosTS}(obj)
}
\arguments{
\item{obj}{an arcosTS object without collective events.}
}
\value{
an arcosTS object
}
\description{
Randomly shift the measurement
}
\details{
Randomly shift/roll the measurement left or right, independently in every trajectory.
Spatial coordinates remain intact.
}
\examples{
cat("no examples")

}
