% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\name{raw_data}
\alias{raw_data}
\title{Download raw ARCOS data (Warning 130+ GB file)}
\usage{
raw_data(key = "WaPo")
}
\arguments{
\item{key}{Key needed to make query successful}
}
\description{
Download raw ARCOS data (Warning 130+ GB file)
}
\examples{
\dontrun{
library(arcos)

the_whole_thing <- raw_data(key="WaPo")

head(the_whole_thing)
}
}
\seealso{
\url{https://www.washingtonpost.com/graphics/2019/investigations/dea-pain-pill-database/#download-resources}

Other raw data: \code{\link{county_raw_fips}},
  \code{\link{county_raw}}, \code{\link{pharmacy_raw}}
}
\concept{raw data}
