% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_clip}
\alias{rcpp_clip}
\title{Clip a numeric vector
Clip a numeric vector between lower and upper bounds.}
\usage{
rcpp_clip(x, a, b)
}
\arguments{
\item{x}{a numeric vector.}

\item{a}{lower bound (double).}

\item{b}{upper bound (double).}
}
\value{
a numeric vector.
}
\description{
Clip a numeric vector
Clip a numeric vector between lower and upper bounds.
}
\examples{
cat("no example")
}
\keyword{internal}
