% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePlotCollEvents2D.R
\name{savePlotColl2D}
\alias{savePlotColl2D}
\alias{savePlotColl2D.arcosTS}
\title{"Save plots of collective events"}
\usage{
savePlotColl2D(
  objTS,
  objColl,
  plotbin = FALSE,
  outdir = ".",
  filecore = "",
  xlim = c(0, 1024),
  ylim = c(0, 1024),
  plotwh = c(3, 3),
  revy = FALSE,
  stretchf = TRUE,
  ggtheme = NULL,
  imtype = c("png", "pdf")
)

\method{savePlotColl2D}{arcosTS}(
  objTS,
  objColl,
  plotbin = FALSE,
  outdir = ".",
  filecore = "",
  xlim = c(0, 1024),
  ylim = c(0, 1024),
  plotwh = c(3, 3),
  revy = TRUE,
  stretchf = TRUE,
  ggtheme = NULL,
  imtype = c("png", "pdf")
)
}
\arguments{
\item{objTS}{an arcosTS object with time series.}

\item{objColl}{an arcosTS object with collective events.}

\item{plotbin}{a boolean, whether dots with binarised measurement should be plotted; default FALSE.}

\item{outdir}{a string with the dierctory to save the folder with frames; default ".".}

\item{filecore}{a string with a prefix for output image files, default ".".}

\item{xlim}{a vector with limits for the x-axis, default c(0, 1024).}

\item{ylim}{a vector with limits for the y-axis, default c(0, 1024).}

\item{plotwh}{a vector with width and height of the output image in inches, default c(3,3).}

\item{revy}{logical whether to reverse the y-axis, default TRUE.}

\item{stretchf}{logical whether to stretch the output sequence to include frames without objects, default TRUE.}

\item{ggtheme}{a ggtheme object with additional style definitions; default NULL.}

\item{imtype}{definition of the output image type, either "png" or "pdf".}
}
\description{
Save plots of collective events
}
\details{
Wrapper for the \code{savePlotCollEvents2D} function.
}
\examples{
cat("no examples")
}
