% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePlotCollEvents2D.R
\name{savePlotCollEvents2D}
\alias{savePlotCollEvents2D}
\title{Save plots of collective events in 2D}
\usage{
savePlotCollEvents2D(
  inDTobj,
  inDTcoll = NULL,
  inDTbin = NULL,
  inDTanno = NULL,
  inCols = list(frame = "frame", x = "x", y = "y", id = "trackID", clid = "clTrackID",
    m = "m", mb = "mb"),
  inDirPath = ".",
  inFileCore = "",
  inXlim = c(0, 1024),
  inYlim = c(0, 1024),
  inPlotWH = c(3, 3),
  inRevY = TRUE,
  inStretchF = TRUE,
  inPtSzMain = 2,
  inPtSzBin = 1,
  inPtSzAnno = 2,
  inGGtheme = NULL,
  inPlotType = c("png", "pdf"),
  inColPal = "RdYlBu"
)
}
\arguments{
\item{inDTobj}{a data.table with time series in long format; required.}

\item{inDTcoll}{a data.table with collective events in long format; e.g. created by trackCollEvents; default NULL.}

\item{inDTbin}{a data.table with a binarised measurement to plot as black dots; default NULL.}

\item{inDTanno}{a data.table with annotations to plot as asterisks.}

\item{inCols}{a list with column names, default list(frame = "frame", x = "x", y = "y", id = "trackID", clid = "clTrackID", m = "m", mb = "mb").}

\item{inDirPath}{a string with the path to save output images, default ".".}

\item{inFileCore}{a string with a prefix for output image files, default ".".}

\item{inXlim}{a vector with limits for the x-axis, default c(0, 1024).}

\item{inYlim}{a vector with limits for the y-axis, default c(0, 1024).}

\item{inPlotWH}{a vector with width and height of the output image in inches, default c(3,3).}

\item{inRevY}{logical, whether to reverse the y-axis, default TRUE.}

\item{inStretchF}{logical, whether to stretch the output sequence to include frames without objects, default TRUE.}

\item{inPtSzMain}{numeric, size of main points, default 1.}

\item{inPtSzBin}{numeric, size of points that indicate binarised measurement, default 0.1.}

\item{inPtSzAnno}{numeric, size of points that indicate annotated data, default 2.}

\item{inGGtheme}{additional ggplot2 theme definitions, default NULL.}

\item{inPlotType}{definition of the output image type, either "png" or "pdf".}

\item{inColPal}{colour palette for the main data, default 'RdYlBu'.}
}
\value{
does not return any value; saves a sequence of images to the specified folder.
}
\description{
Save subsequent frames as PNG or PDF files with optional collective events overlaid as convex hull polygons
over the original data. Optionally, a binarised measurement can be plotted as black dots and
arbitrary annotations as asterisks.
}
\examples{
cat("no example")
}
\keyword{internal}
