% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selCollEvents.R
\name{selCollEvents}
\alias{selCollEvents}
\title{Wrapper for filtering collective events}
\usage{
selCollEvents(
  inDTcoll,
  inCols = list(frame = "frame", id = "trackID", collid = "clTrackID"),
  inCollDur,
  inCollTotSz
)
}
\arguments{
\item{inDTcoll}{a data.table with collective events in the long format produced by the \code{trackCollEvents} function. Consists of 3 columns: integer frame number, object id, id's of collective events.}

\item{inCols}{a list with column names, \code{list(frame = , id = , collid = )}, that correspond to the integer frame number, position, object id and id of collective events, respectively.}

\item{inCollDur}{a vector with 2 integers that correspond to the minimum and maximum duration of clusters in frame numbers.}

\item{inCollTotSz}{a vector with 2 integers that correspond to the minimum and maximum total size of clusters in frame numbers. The total size is the number of unique objects involved in a collective event throughout its entire duration.}
}
\value{
a data.table as the \code{inDTcoll} with collective events within provided duration and size bounds.
}
\description{
Filter collective events calculated by the \code{trackCollEvents} function by size and duration.
}
\examples{
cat("no examples")
}
