% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shifter}
\alias{shifter}
\title{Vector shifter}
\usage{
shifter(x, n = 1)
}
\arguments{
\item{x}{a numeric vector}

\item{n}{an integer, the number of indices to shift the input vector. Positive, shifts left; negative, shifts right.}
}
\value{
a numeric vector
}
\description{
Shift a vector left or right by a number of spaces.
From: https://stackoverflow.com/a/30542172/1898713
library(SOfun); shifter(a)#'
}
\examples{
library(ARCOS)
ARCOS:::shifter(1:10, 2)
}
