% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffBlockTrack.R
\name{shuffBlockTrack}
\alias{shuffBlockTrack}
\alias{shuffBlockTrack.arcosTS}
\title{"Shuffle blocks of the measurement per track"}
\usage{
shuffBlockTrack(obj)

\method{shuffBlockTrack}{arcosTS}(obj)
}
\arguments{
\item{obj}{an arcosTS object without collective events.}
}
\value{
an arcosTS object
}
\description{
Shuffle blocks of binarised measurement per track
}
\details{
Shuffle blocks of 0s & 1s independently for every track.
Spatial coordinates remain intact.
}
\examples{
cat("no examples")

}
