% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shuffBlockVec}
\alias{shuffBlockVec}
\title{Block shuffle a binary vector}
\usage{
shuffBlockVec(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector
}
\description{
Randomly shuffle runs of 0s & 1s in a vector. Do not maintain the alternating order of 0s & 1s.
We assume that the vector consists of 0s & 1s.
}
\examples{
library(ARCOS)
set.seed(7)
x <- round(runif(20))
ARCOS:::shuffBlockVec(x)
}
