% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shuffBlockVecAlt}
\alias{shuffBlockVecAlt}
\title{Block shuffle a binary vector}
\usage{
shuffBlockVecAlt(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector
}
\description{
Randomly shuffle runs of 0s & 1s in a vector but maintain their alternating order,
i.e. there'll never be joint runs of 0s or 1s from the original sequence.
We assume that the vector consists of 0s & 1s.
}
\examples{
library(ARCOS)
set.seed(7)
x <- round(runif(20))
ARCOS:::shuffBlockVecAlt(x)
}
