% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffCoord.R
\name{shuffCoord}
\alias{shuffCoord}
\alias{shuffCoord.arcosTS}
\title{"Shuffle coordinates"}
\usage{
shuffCoord(obj)

\method{shuffCoord}{arcosTS}(obj)
}
\arguments{
\item{obj}{an arcosTS object without collective events.}
}
\value{
an arcosTS object
}
\description{
Shuffle coordinates
}
\details{
Shuffle spatial coordinates of all tracks in the dataset.
Time points and the measurement remain intact.

First, initial positions of all tracks are shuffled between the objects.
Then, original displacements are used to calculate new coordinates with
respect to the new origin. This approach of coordinate shuffling is
useful for bootstrapping of data where objects don't change spatial
coordinates over time. Otherwise, the trajectories with
new initial spatial coordinates go outside of the FOV over time.
}
\examples{
cat("no examples")

}
