% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffMeasFrame.R
\name{shuffMeasFrame}
\alias{shuffMeasFrame}
\alias{shuffMeasFrame.arcosTS}
\title{"Shuffle the measurement per frame"}
\usage{
shuffMeasFrame(obj)

\method{shuffMeasFrame}{arcosTS}(obj)
}
\arguments{
\item{obj}{an arcosTS object without collective events.}
}
\value{
an arcosTS object
}
\description{
Shuffle the measurement per frame
}
\details{
Shuffle the measurement independently in every frame.
Spatial coordinates remain intact.
}
\examples{
cat("no examples")

}
