% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffMeasTrack.R
\name{shuffMeasTrack}
\alias{shuffMeasTrack}
\alias{shuffMeasTrack.arcosTS}
\title{"Shuffle the measurement per track"}
\usage{
shuffMeasTrack(obj)

\method{shuffMeasTrack}{arcosTS}(obj)
}
\arguments{
\item{obj}{an arcosTS object without collective events.}
}
\value{
an arcosTS object
}
\description{
Shuffle the measurement per track
}
\details{
Shuffle time points in the measurement independently for every track.
Spatial coordinates remain intact.
}
\examples{
cat("no examples")

}
