% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackCollEvents.R
\name{trackColl}
\alias{trackColl}
\alias{trackColl.arcosTS}
\title{"Track collective events"}
\usage{
trackColl(obj, eps = 1, minClSz = 1L, nPrev = 1L, deb = FALSE)

\method{trackColl}{arcosTS}(obj, eps = 1, minClSz = 1L, nPrev = 1L, deb = FALSE)
}
\arguments{
\item{obj}{an arcosTS object.}

\item{eps}{a numeric, sets the search radius; default 1.}

\item{minClSz}{an integer, minimum cluster size; default 1L.}

\item{nPrev}{an integer, number of previous frames to link; default 1L.}

\item{deb}{boolean, additional debug output; default FALSE,}
}
\value{
an arcosTS object
}
\description{
Track collective events
}
\details{
Wrapper for the \code{trackCollEvents} function.
}
\examples{
library(ARCOS)
library(data.table)
ts = data.table(frame = c(1, 1, 2, 2, 2, 3, 3, 3, 3, 4, 4, 4, 5, 5),
                objid = c(1, 2, 1, 2, 3, 1, 2, 3, 4, 1, 2, 4, 1, 4),
                x = c(1, 3, 1.2, 2.5, 3.5, 0.9, 2.6, 2.9, 3.2, 1.1, 2.8, 3.1, 1, 3))

arcosTS(ts,
        colPos = "x",
        col = list(Frame = "frame",
                   IDobj = "objid",
                   RT = NULL,
                   IDcoll = NULL),
        interVal = 1.,
        interType = "fixed")

tc = trackColl(ts)
}
