% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackCollBoot.R
\name{trackCollBoot}
\alias{trackCollBoot}
\alias{trackCollBoot.arcosTS}
\title{"Track collective events with bootstrapping randomisation"}
\usage{
trackCollBoot(
  obj,
  nboot = 100L,
  ncores = 2L,
  method = c("shuffCoord", "randShiftMeas", "shuffMeasTrack", "shuffMeasFrame",
    "shuffBlockTrack"),
  eps = 1,
  minClSz = 1L,
  nPrev = 1L,
  epsPrev = NULL,
  colldurlim = c(1, Inf),
  colltotszlim = c(1, Inf),
  deb = FALSE
)

\method{trackCollBoot}{arcosTS}(
  obj,
  nboot = 100L,
  ncores = 2L,
  method = c("shuffCoord", "randShiftMeas", "shuffMeasTrack", "shuffMeasFrame",
    "shuffBlockTrack"),
  eps = 1,
  minClSz = 1L,
  nPrev = 1L,
  epsPrev = NULL,
  colldurlim = c(1, Inf),
  colltotszlim = c(1, Inf),
  deb = FALSE
)
}
\arguments{
\item{obj}{an arcosTS object.}

\item{nboot}{an integer, number of bootstrap iterations; default 100.}

\item{ncores}{an integer, number of parallel cores; default 2.}

\item{method}{either of the five bootstrapping methods: shuffCoord', 'randShiftMeas', 'shuffMeasTrack', 'shuffMeasFrame', 'shuffBlockTrack'.}

\item{eps}{a numeric, sets the search radius for spatial clustering with dbscan; default 1.}

\item{minClSz}{an integer, minimum cluster size for dbscan; default 1L.}

\item{nPrev}{an integer, number of previous frames to link; default 1L.}

\item{epsPrev}{a float with the search radius for linking clusters between frames, default NULL means that epsPrev = eps.}

\item{colldurlim}{a two-element vector with limits for filtering cluster duration; default c(1, Inf).}

\item{colltotszlim}{a two-element vector with limits for filtering cluster size; default c(1, Inf).}

\item{deb}{boolean, additional debug output; default FALSE.}
}
\value{
a data.table with filtered stats of collective events at every bootstrapping iteration
}
\description{
Track collective events with bootstrapping randomisation
}
\details{
Identifies collective events using one of five randomisation methods.
}
\examples{
cat("No example\n")
}
