% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcos-package.R
\docType{package}
\name{arcos}
\alias{arcos}
\title{\code{arcos} package}
\description{
Download and use data from ARCOS, raw or summarized by the Washington Post, in R
}
\author{
Andrew Ba Tran (@abtran)

Steven Rich (@dataeditor)

Aaron Williams (@aboutaaron)
}
