% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcosTS.R
\name{arcosTS}
\alias{arcosTS}
\title{Create an arcosTS class}
\usage{
arcosTS(
  dt,
  colPos = "x",
  colMeas = NULL,
  col = list(Frame = "f", IDobj = "id", RT = NULL, IDcoll = NULL),
  interVal = NULL,
  interType = c("fixed")
)
}
\arguments{
\item{dt}{a data.table with time series in the long format.}

\item{colPos}{a vector with names of positional columns; default "x".}

\item{colMeas}{a string with the column name of the measurement; default NULL.}

\item{col}{a list with names of other columns, i.e. list(Frame = , IDobj = , RT = , IDcoll = ), with names of the frame, object id, real time, and collective id columns.}

\item{interVal}{a numeric with the interval length.}

\item{interType}{a string to designate whether the time series has fixed or variable intervals; possible values fixed or var.}
}
\value{
an arcosTS object.
}
\description{
Creates an arcosTS object from time series data in long format stored in a `data.table`. Assigns relevant column names and data parameters.
}
\examples{
library(ARCOS)
library(data.table)
dts = arcosTS(dt = data.table(frame = c(1, 1, 2, 2, 2, 3, 3, 3, 3, 4, 4, 4, 5, 5),
                              id = c(1, 2, 1, 2, 3, 1, 2, 3, 4, 1, 2, 4, 1, 4),
                              x = c(1, 3, 1.2, 2.5, 3.5, 0.9, 2.6, 2.9, 3.2, 1.1, 2.8, 3.1, 1, 3)),
              colPos = "x",
              col = list(Frame = "frame",
                         IDobj = "id"),
              interVal = 1,
              interType = "fixed")
}
