% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPvalFromBS.R
\name{calcPvalFromBS}
\alias{calcPvalFromBS}
\title{Calculate p-value from bootstrapping}
\usage{
calcPvalFromBS(
  testStat,
  bsStat,
  corrected = FALSE,
  alternative = c("greater", "less")
)
}
\arguments{
\item{testStat}{a numeric, test statistic.}

\item{bsStat}{a numeric vector with statistic from bootstrapping.}

\item{corrected}{a logical; if TRUE, calculates a p-value that accounts for the finite sample size; default FALSE.}

\item{alternative}{either of "less" or "greater"; default "greater"}
}
\value{
a scalar with the p-value
}
\description{
The calculation of the p-value that accounts for the finite sample size uses a formula from Davison and Hinkley (1997), Bootstrap Methods and their Application, chapter 4, p. 161.
}
\examples{
library(ARCOS)
ARCOS::calcPvalFromBS(2, rnorm(1000))
}
