% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsCollEvents.R
\name{calcStatsColl}
\alias{calcStatsColl}
\alias{calcStatsColl.arcosTS}
\title{"Select collective events"}
\usage{
calcStatsColl(obj)

\method{calcStatsColl}{arcosTS}(obj)
}
\arguments{
\item{obj}{an arcosTS object.}

\item{eps}{a numeric, sets the search radius.}

\item{minpts}{an integer, minimum cluster size.}

\item{nprev}{an integer, number of previous frames to link.}
}
\value{
an arcosTS object
}
\description{
Select collective events
}
\details{
Wrapper for the \code{selCollEvents} function.
}
\examples{
cat("no examples")
}
