% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsCollEvents.R
\name{calcStatsColl}
\alias{calcStatsColl}
\alias{calcStatsColl.arcosTS}
\title{"Stats of collective events"}
\usage{
calcStatsColl(obj)

\method{calcStatsColl}{arcosTS}(obj)
}
\arguments{
\item{obj}{an arcosTS object.}
}
\value{
an arcosTS object
}
\description{
Stats of collective events
}
\details{
Wrapper for the \code{calcStatsCollEvents} function.
}
\examples{
cat("no examples")
}
