% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatsCollEvents.R
\name{calcStatsCollEvents}
\alias{calcStatsCollEvents}
\title{Wrapper for calculating stats of collective events}
\usage{
calcStatsCollEvents(
  inDTcoll,
  inCols = list(frame = "frame", id = "trackID", collid = "clTrackID")
)
}
\arguments{
\item{inDTcoll}{a data.table with collective events in the long format produced by the \code{trackCollEvents} function. Consists of 3 columns: integer frame number, object id, id's of collective events.}

\item{inCols}{a list with column names, \code{list(frame = , id = , collid = )}, that correspond to the integer frame number, position, object id and id of collective events, respectively.}
}
\value{
a data.table as the \code{inDTcoll} with aggregated stats of collective events.
}
\description{
Stats of collective events calculated by the \code{trackCollEvents} function by size and duration.
}
\examples{
cat("no examples")
}
