% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genSynth2D}
\alias{genSynth2D}
\title{A single synthetic collective event in 2D}
\usage{
genSynth2D(inSeed = NULL)
}
\arguments{
\item{inSeed}{an integer with the seed for the random number generator, default NULL.}
}
\value{
an arcosTS object
}
\description{
81 objects in 2D in 8 time frames. X/Y positions have an additional small Gaussian noise.
}
\examples{
library(ARCOS)
dts = genSynth2D()
}
