% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataFromFile.R
\name{loadDataFromFile}
\alias{loadDataFromFile}
\title{Title}
\usage{
loadDataFromFile(
  fname,
  colPos = "x",
  colMeas = NULL,
  col = list(Frame = "f", IDobj = "id", RT = NULL, IDcoll = NULL),
  interVal = 1L,
  interType = c("fixed"),
  ...
)
}
\arguments{
\item{fname}{a string with the file name path.}

\item{colPos}{a vector with names of positional columns; default "x".}

\item{colMeas}{a string with the column name of the measurement; default NULL.}

\item{col}{a list with names of other columns, i.e. list(Frame = , IDobj = , RT = , IDcoll = ), with names of the frame, object id, real time, and collective id columns.}

\item{interVal}{a numeric with the interval length.}

\item{interType}{a string to designate whether the time series has fixed or variable intervals; possible values fixed or var.}

\item{...}{additional parameters passed to data.table::fread.}
}
\value{
an arcosTS object.
}
\description{
Title
}
\examples{
cat("no examples")
}
