% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataFromFile.R
\name{loadDataFromFile}
\alias{loadDataFromFile}
\title{Load data from file}
\usage{
loadDataFromFile(
  fname,
  colPos = "x",
  colFrame = "f",
  colIDobj = "id",
  colIDcoll = NULL,
  colMeas = NULL,
  colMeasResc = NULL,
  colMeasBin = NULL,
  colBootIter = NULL,
  colRT = NULL,
  interVal = 1L,
  interType = c("fixed"),
  fromBin = FALSE,
  fromColl = FALSE,
  fromBoot = FALSE,
  ...
)
}
\arguments{
\item{fname}{a string, file name path.}

\item{colPos}{a vector, names of positional columns; default "x".}

\item{colFrame}{a string, column name with frame numbers; default "f".}

\item{colIDobj}{a string, column name with object IDs; default "id".}

\item{colIDcoll}{a string, column name with IDs of collective events; default NULL.}

\item{colMeas}{a string, column name of the measurement; default NULL.}

\item{colMeasResc}{a string, column name of the rescaled measurement; default NULL.}

\item{colMeasBin}{a string, column name of the binarised measurement; default NULL.}

\item{colBootIter}{a string, column name with bootstrap iterations; default NULL.}

\item{colRT}{a string, column name with real time; default NULL (currently unused).}

\item{interVal}{a numeric with the interval length (currently unused).}

\item{interType}{a string to designate whether the time series has fixed or variable intervals; possible values fixed or var (currently unused).}

\item{fromBin}{logical, whether the output has been passed by binarisation, relevant for plotting of binarised activity; default FALSE.}

\item{fromColl}{logical, whether the output comes from collective event identification, relevant for plotting of collective events; default FALSE.}

\item{fromBoot}{logical, whether the output comes from bootstrapping, relevant for calculating stats of collective events; default FALSE.}

\item{...}{additional parameters passed to data.table::fread.}
}
\value{
an arcosTS object.
}
\description{
Loads data from a file and creates an arcosTS object based on provided parameters.
Data is loaded using data.table::fread function, thus all file formats recognised
by fread are admissible, including gzip and bz2 compressions.
}
\examples{
cat("no examples")
}
