% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBinMeas.R
\name{plotBinMeas}
\alias{plotBinMeas}
\alias{plotBinMeas.arcosTS}
\title{"Visualise measurement binarisation"}
\usage{
plotBinMeas(obj, ntraj = 16L, measfac = 1)

\method{plotBinMeas}{arcosTS}(obj, ntraj = 16L, measfac = 1)
}
\arguments{
\item{obj}{an arcosTS object.}

\item{ntraj}{an integer with the number of random trajectories to plot; default 16L.}

\item{measfac}{a numeric with a rescaling factor for the measurement for visualisation purposes; default 1.}
}
\value{
a ggplot2 object.
}
\description{
Visualise measurement binarisation
}
\examples{
cat("no examples")
}
