% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMeasBin.R
\name{plotBinMeas}
\alias{plotBinMeas}
\alias{plotBinMeas.arcosTS}
\title{"Visualise measurement binarisation"}
\usage{
plotBinMeas(obj, ntraj = 16, measfac = 1)

\method{plotBinMeas}{arcosTS}(obj, ntraj = 16, measfac = 1)
}
\arguments{
\item{obj}{an arcosTS object.}
}
\value{
a ggplot2 object.
}
\description{
Visualise measurement binarisation
}
\examples{
cat("no examples")
}
