% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNoodle2D.R
\name{plotNoodle2D}
\alias{plotNoodle2D}
\alias{plotNoodle2D.arcosTS}
\title{"Plot collective events"}
\usage{
plotNoodle2D(objTS, xfac = 1, yfac = 1, tfreq = 1, pos = 1L)

\method{plotNoodle2D}{arcosTS}(objTS, xfac = 1, yfac = 1, tfreq = 1, pos = 1L)
}
\arguments{
\item{xfac}{a numeric, multiplication factor for the x-axis; default 1.}

\item{yfac}{a numeric, multiplication factor for the y-axis; default 1.}

\item{tfreq}{a numeric, frequency of the temporal column; default 1.}

\item{pos}{an integer, index of the coordinate to plot in the position vector; default 1L.}

\item{obj}{an arcosTS object.}
}
\value{
a ggplot2 object.
}
\description{
Plot tracks from an arcosTS object
}
\details{
Plot tracks in 2D.
}
\examples{
cat("no examples")
}
