% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selCollEvents.R
\name{selColl}
\alias{selColl}
\alias{selColl.arcosTS}
\title{Select collective events}
\usage{
selColl(obj, colldur = c(1, Inf), colltotsz = c(1, Inf))

\method{selColl}{arcosTS}(obj, colldur = c(1, Inf), colltotsz = c(1, Inf))
}
\arguments{
\item{obj}{an arcosTS object.}

\item{eps}{a numeric, sets the search radius.}

\item{minpts}{an integer, minimum cluster size.}

\item{nprev}{an integer, number of previous frames to link.}
}
\value{
an arcosTS object
}
\description{
Wrapper for the \code{selCollEvents} function.
}
\examples{
dt = data.table(frame = c(1, 1, 2, 2, 2, 3, 3, 3, 3, 4, 4, 4, 5, 5),
                objid = c(1, 2, 1, 2, 3, 1, 2, 3, 4, 1, 2, 4, 1, 4),
                x = c(1, 3, 1.2, 2.5, 3.5, 0.9, 2.6, 2.9, 3.2, 1.1, 2.8, 3.1, 1, 3))

ts = arcosTS(dt,
             posCols = c("x"),
             cols = list(frame = "frame",
             id = "objid"),
             interVal = 1.,
             interType = "fixed")
ts = trackColl(ts)
ts
}
