% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackLen.R
\name{selTrackLen}
\alias{selTrackLen}
\alias{selTrackLen.arcosTS}
\title{Select track lengths}
\usage{
selTrackLen(obj, lenmin = NULL, lenmax = NULL)

\method{selTrackLen}{arcosTS}(obj, lenmin = NULL, lenmax = NULL)
}
\arguments{
\item{obj}{an arcosTS object.}

\item{lenmin}{an integer, minimum track length; default NULL.}

\item{lenmax}{an integer, maximum track length; default NULL.}
}
\value{
an arcosTS object.
}
\description{
Select track lengths
}
\examples{
cat("no examples")
}
