% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffBlockTrack.R
\name{shuffBlockTrack}
\alias{shuffBlockTrack}
\alias{shuffBlockTrack.arcosTS}
\title{"Shuffle blocks of the measurement per track"}
\usage{
shuffBlockTrack(obj, alt = FALSE)

\method{shuffBlockTrack}{arcosTS}(obj, alt = FALSE)
}
\arguments{
\item{obj}{an arcosTS object without collective events.}

\item{alt}{logical, whether to maintain the alternating order of 0s & 1s; default FALSE.}
}
\value{
an arcosTS object
}
\description{
Shuffle blocks of binarised measurement per track
}
\details{
Shuffle blocks of 0s & 1s independently for every track.
Spatial coordinates remain intact.
}
\examples{
cat("no examples")

}
