% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffCoord.R
\name{shuffCoord}
\alias{shuffCoord}
\alias{shuffCoord.arcosTS}
\title{"Shuffle coordinates"}
\usage{
shuffCoord(obj)

\method{shuffCoord}{arcosTS}(obj)
}
\arguments{
\item{obj}{an arcosTS object without collective events.}
}
\value{
an arcosTS object
}
\description{
Shuffle coordinates
}
\details{
Shuffle spatial coordinates of all tracks in the dataset.
The time points and the measurement remain intact.

First, initial positions of every track are shuffled between all objects.
Then original displacements are used to calculate new coordinates with
respect to the new origin.
}
\examples{
cat("no examples")

}
