% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackCollEvents.R
\name{trackCollEvents}
\alias{trackCollEvents}
\title{Track collective events}
\usage{
trackCollEvents(
  inDT,
  inEps = 1,
  inMinPts = 1L,
  inNprev = 1,
  inCols = list(frame = "time", x = "x", y = NULL, z = NULL, trackID = "trackID",
    clTrackID = "clTrackID"),
  DEB = T
)
}
\arguments{
\item{DEB}{}
}
\value{

}
\description{
Track collective events
}
\examples{

}
